Refactor app to better separate the logic (#26)
This removes the read and write closures and makes them static member
functions instead. This makes it clearer where the code is and makes it
more scalable when we have more methods.

This also adds in custom grpc endpoint registrations that handle the
grpc-specific format: always a POST request, method formed of components
and using the grpc adapters for the endpoint function.