Add Value type and map manipulation methods that use it
This allows us to store fields with the value, like create revision.

The create revision can't be set inside the transaction that creates the
value since we don't know the id, but the next time we access it we can
fill it in, either just in the response to the user or in a new KV when
overwriting it.